Fixed some python 3.5 bugs

+++ glumpy/graphics/collections/agg_path_collection.py
@@ -187,11 +187,11 @@ class AggPathCollection(Collection):
         n = itemsize
         if closed:
             I = np.resize(np.array([0,1,2, 1,2,3], dtype=np.uint32),n*2*3)
-            I += np.repeat( 4*np.arange(n), 6)
+            I += np.repeat( 4*np.arange(n, dtype=np.uint32), 6)
             I[-6:] = 4*n-6,4*n-5,0,4*n-5,0,1
         else:
             I = np.resize(np.array([0,1,2, 1,2,3], dtype=np.uint32),(n-1)*2*3)
-            I += np.repeat( 4*np.arange(n-1), 6)
+            I += np.repeat( 4*np.arange(n-1, dtype=np.uint32), 6)
         I = I.ravel()
 
         # Uniforms
